Add files via upload
Yesterday problem has been solved.
JavaScript function is moved to button onClicked. 

It requires rethinking whether it is worth creating JS functions when creating in QtQuick. 
At the moment it seems that the C++ backend does not impose such restrictions as JS. 
The advantage of functions in JS is that in smaller projects they can exist in the same file as the QML front-end.